Don't use "in_utf" macro in terminal.c
It has a hidden dependancy on the line codepage that forces UTF-8 mode if
the codepage is UTF-8, I don't want that. I do still want the "utf" flag
to notice this though, only on startup is fine.

Note: The ldisc file still uses it because if we are stating that the
line is UTF-8 we probably don't want the keyboard translation to notice
that the host has snuck over to iso2022 mode so it can use the VT100
alternate character set.